+++ CafeMaestro/Views/Settings/AboutView.xaml.cs
@@ -1,0 +1,6 @@
+namespace CafeMaestro.Views.Settings;
+
+public partial class AboutView : ContentView
+{
+    public AboutView() => InitializeComponent();
+}